PRACTICA5/1/api/requirements.txt
flask==2.0.1
flask-restful==0.3.9
flask-sqlalchemy==1.4.54
flask-cors==3.0.10
pymysql==1.0.2
python-dotenv==0.19.0
werkzeug==2.0.1